+++ src/lava/proc/event_temp/flattening.py
